replace v2 component
