+++ jest.config.json
@@ -1,5 +1,5 @@
 {
   "collectCoverage": true,
   "testEnvironment": "jest-environment-jsdom",
-  "testRegex": ".(test|spec).tsx?$"
+  "testRegex": ".test.tsx?$"
 }
