+++ src/pin-field/index.ts
@@ -1,2 +1,1 @@
 export * from "./pin-field";
-export * from "./pin-field.types";
