+++ src/pin-field/pin-field.e2e.ts
@@ -5,7 +5,7 @@ describe("PIN Field", () => {
   beforeEach(() => {
     cy.visit("/iframe.html", {
       qs: {
-        id: "pinfieldv2--default",
+        id: "pinfield--default",
         viewMode: "story",
       },
     });
