--- src/polyfills/index.ts
@@ -1,1 +1,0 @@
-export * from "./keyboard-evt";
