+++ src/utils/utils.test.ts
@@ -1,4 +1,4 @@
-import { noop, range, omit } from "./utils";
+import { noop, range } from "./utils";
 
 test("noop", () => {
   expect(noop()).toEqual(undefined);
@@ -9,11 +9,4 @@ test("range", () => {
   expect(range(0, 3)).toEqual([0, 1, 2]);
   expect(range(3, 0)).toEqual([]);
   expect(range(3, 3)).toEqual([3, 4, 5]);
-});
-
-test("omit", () => {
-  expect(omit([], { a: 0, b: 1 })).toEqual({ a: 0, b: 1 });
-  expect(omit(["a"], { a: 0, b: 1 })).toEqual({ b: 1 });
-  expect(omit(["b"], { a: 0, b: 1 })).toEqual({ a: 0 });
-  expect(omit(["a", "b"], { a: 0, b: 1 })).toEqual({});
 });
