+++ vite.config.js
@@ -1,9 +1,9 @@
 import { resolve } from "path";
-import type { UserConfig } from "vite";
+import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import dts from "vite-plugin-dts";
 
-export default {
+export default defineConfig({
   plugins: [react(), dts()],
   build: {
     lib: {
@@ -24,4 +24,4 @@ export default {
       },
     },
   },
-} satisfies UserConfig;
+});
